+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [2.1.0] - 2018-12-12
+- Fix processing status updates.
+- Add source information filters.
+
 ## [2.0.0] - 2018-05-14
 - Switched to PHP namespaces.
 
@@ -22,7 +26,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-01-20
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/2.0.0...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/2.1.0...HEAD
+[2.1.0]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/1.0.3...2.0.0
 [1.0.3]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/wp-pay-extensions/event-espresso-legacy/compare/1.0.1...1.0.2